FEATURE: use new link editor in inspector
