+++ packages/neos-ui-editors/src/Library/LinkInput.css
@@ -1,0 +1,62 @@
+.linkInput__wrap {
+    display: flex;
+    align-items: stretch;
+}
+
+.linkInput__optionWrapper {
+    flex-grow: 1;
+    flex-shrink: 1;
+    /* see https://css-tricks.com/flexbox-truncated-text/ */
+    min-width: 0;
+}
+
+.linkInput__optionsPanelDouble {
+    display: flex;
+    flex-basis: 100%;
+}
+
+.linkInput__optionsPanel {
+    padding-top: var(--spacing-Full);
+    display: flex;
+    flex-wrap: wrap;
+}
+
+.linkInput__optionsPanelItem {
+    flex-basis: 50%;
+    padding: var(--spacing-Half);
+}
+
+.linkInput__optionsPanelLabel {
+    margin-bottom: var(--spacing-Quarter);
+}
+
+.linkInput__innerButton {
+    height: auto;
+}
+
+.linkInput__loader {
+    margin: var(--spacing-Quarter);
+}
+
+.linkInput__item {
+    padding: var(--spacing-Half);
+    box-sizing: content-box;
+    height: calc(var(--spacing-GoldenUnit) * 1.5);
+    line-height: 20px;
+}
+
+.linkInput__link {
+    color: var(--colors-ContrastBright);
+    display: block;
+    font-size: 12px;
+}
+
+.linkInput__image {
+    height: calc(var(--spacing-GoldenUnit) * 1.5);
+    width: auto;
+    float: right;
+}
+
+.linkInput__link--active {
+    color: #fff;
+}
